Feature: Improve tag selection behavior

+++ SearchPubChem/Compounds/Controllers/CompoundTagViewController.swift
@@ -29,6 +29,7 @@ class CompoundTagViewController: UIViewController {
     var compound: Compound!
     private var tagsAttachedToCompound = Set<CompoundTag>()
     private var sellectedCells = Set<IndexPath>()
+    private var lastSelectedCell: IndexPath?
     
     var dataController: DataController!
     var fetchedResultsController: NSFetchedResultsController<CompoundTag>! {
@@ -129,6 +130,9 @@ class CompoundTagViewController: UIViewController {
             newTag.compoundCount = 1
             newTag.name = newTagTextField.text
             
+            tagsAttachedToCompound.insert(newTag)
+            setTagsLabel()
+            
             if let tags = compound.tags, tags.count > 0 {
                 tags.adding(newTag)
             } else {
@@ -141,24 +145,22 @@ class CompoundTagViewController: UIViewController {
                 NSLog("Error while saving in iPadCompoundTagViewController.addNewTag(:)")
             }
             
-            tagsAttachedToCompound.insert(newTag)
-            setTagsLabel()
         } else {
             print("New tag is not given")
         }
     }
     
     @IBAction func deleteTag(_ sender: UIButton) {
-        for indexPath in sellectedCells {
+        if let indexPath = lastSelectedCell {
             let tag = fetchedResultsController.object(at: indexPath)
             dataController.viewContext.delete(tag)
-        }
-        
-        do {
-            try dataController.viewContext.save()
-            NSLog("Saved in iPadCompoundTagViewController.deleteTags(:)")
-        } catch {
-            NSLog("Error while saving in iPadCompoundTagViewController.deleteTags(:)")
+            
+            do {
+                try dataController.viewContext.save()
+                NSLog("Saved in iPadCompoundTagViewController.deleteTags(:)")
+            } catch {
+                NSLog("Error while saving in iPadCompoundTagViewController.deleteTags(:)")
+            }
         }
     }
     
@@ -202,6 +204,10 @@ extension CompoundTagViewController: UICollectionViewDelegate, UICollectionViewD
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let tag = fetchedResultsController.object(at: indexPath)
+        if tagsAttachedToCompound.contains(tag) && !sellectedCells.contains(indexPath) {
+            sellectedCells.insert(indexPath)
+        }
+        
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: collectionViewCellIdentifier, for: indexPath) as! iPadCompoundTagCollectionViewCell
         
         cell.nameLabel.text = tag.name
@@ -226,10 +232,22 @@ extension CompoundTagViewController: UICollectionViewDelegate, UICollectionViewD
 
         setTagsLabel()
         
+        lastSelectedCell = lastSelectedCell == indexPath ? nil : indexPath
+        
         return true
     }
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        if let cell = collectionView.cellForItem(at: indexPath) as? iPadCompoundTagCollectionViewCell {
+            cell.containerView.backgroundColor = sellectedCells.contains(indexPath) ? .cyan : .white
+            
+            if lastSelectedCell == indexPath {
+                cell.containerView.backgroundColor = .red
+            }
+        }
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
         if let cell = collectionView.cellForItem(at: indexPath) as? iPadCompoundTagCollectionViewCell {
             cell.containerView.backgroundColor = sellectedCells.contains(indexPath) ? .cyan : .white
         }
